Multi agent, and somewhat Unity's ML-Toolkit support
The main change is making sure that MADDPG works.
It works the example isn't be the best to check. For now, the example uses Unity's modified tennis example. In the future there will better integration with the ML-Tookits, for now we added a custom env runner and gym-like wrapper.
The custom MultiAgentEnvRunner will need plenty of updates depending on required tasks. Currently it assume that the task is coop with total score maximization.