Isolate Rainbow from DQN & small bug fixes
Although Rainbow is a specific combination of DQN extensions and could be/was treated as DQN with a particular config, it was a mental burden. Potential lesson for future is to have a vanilla flavour for each popular agent but maybe have a separate class that is easy extensible. One would be for teaching and testing reference, the other would be for tinkering.

There was also a bug in NormalNoise policy. I assumed it was my pytorch version but it was actually using univariate Normal distribution with parameters for multi variate.

Another was passing twice the same keyword params in critic dual nets.