Migrate agents to spaces (n-dim) from sizes (1-dim) (#14)
Added DataSpace type which (as the name suggests) describes space of data. It's similar to gym's spaces and can be used as its wrapper. This is requirement to move with agents into multi dimensional input and output spaces with consistent interface.

Most agents have been tested and made sure that nothing has been changed from performance. Some, however, will need to be extra tested to make sure that algorithms work properly.

This PR also drops support for Python3.6. We're using a few features that are only in 3.8+ like `cached_property` or recurrent annotations. This bumps the version to v0.2.0.